pyproject.toml
[tool.poetry]
name = "lab-flask"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "lab_flask"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
flask-restful = "^0.3.10"
flask-sqlalchemy = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
